refactor : 상품 디테일 리렌더링 수정
